requirements.txt
langchain>=0.1.0
langchain-community>=0.2.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pinecone>=7.0.0
openai>=1.0.0

# Web UI dependencies
streamlit>=1.28.0
